backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==18.2.0
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
cryptography==45.0.7
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pdflatex==0.1.3
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.6.1
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
openai==1.54.4
tiktoken==0.8.0

# Database
SQLAlchemy==2.0.41
alembic==1.16.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

setuptools>=70.0.0
watchdog==4.0.1
PyJWT==2.8.0
razorpay==1.4.2

# Redis and Celery for Phase 8 - Workers & Queue System
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Additional dependencies for Phase 8
aioredis==2.0.1
kombu==5.3.4

# Test dependencies
pytest-asyncio==1.2.0
httpx==0.27.0
pgvector==0.3.6
